Moved some match starting code the abstract class IGameBoard
They seem better suited for the base class IGameBoard because they are
common to all gameboards. I also added a return type of bool because I
will eventually add an "Are you sure?" dialogue and I want the caller of
these functions to be able to know if the match was successfully
ended/started.