Added comments/useful mutators/RowCol struct.
Added RowCol to represent a piece on the board and whether or not that
is a 'valid'. Right now valid just offers a way for the code to know if
this position is "good" or "bad". e.g. I mark m_selected's RowCol as
false whenever the current player changes to signify that it is not the
position the new player selected.